Kissimmee watershed land use: https://www.lake.wateratlas.usf.edu/watershed/geography.asp?wshedid=1&wbodyatlas=watershed#use
I took the information from the website and made a table in R (based on the online table shown below). I think this is a lot of what they wanted in terms of land use information. I just need to clarify timeframe with them. Do they want the most recent year of land use/cover? Or all years available?